+++ frontend/src/pages/FavoritesPage.vue
@@ -22,6 +22,8 @@ const customPageSize = ref('')
 
 onMounted(async () => {
   gallery.viewMode = 'favorites'
+  gallery.continueWatching = false
+  gallery.tagFilter = ''
   gallery.category = null
   gallery.folder = null
   gallery.page = 1